refactor(sanity): move release filtering code to reusable `useFilteredReleases` hook (#7907)
