+++ packages/sanity/src/structure/panes/document/documentPanel/header/perspective/DocumentPerspectiveList.tsx
@@ -13,10 +13,8 @@ import {
   type ReleaseDocument,
   Translate,
   useActiveReleases,
-  useArchivedReleases,
   useDateTimeFormat,
   type UseDateTimeFormatOptions,
-  useDocumentVersions,
   usePerspective,
   useSchema,
   useSetPerspective,
@@ -26,13 +24,8 @@ import {
 import {usePaneRouter} from 'sanity/structure'
 
 import {isLiveEditEnabled} from '../../../../../components/paneItem/helpers'
+import {useFilteredReleases} from '../../../../../hooks/useFilteredReleases'
 import {useDocumentPane} from '../../../useDocumentPane'
-
-type FilterReleases = {
-  notCurrentReleases: ReleaseDocument[]
-  currentReleases: ReleaseDocument[]
-  inCreation: ReleaseDocument | null
-}
 
 const TooltipContent = ({release}: {release: ReleaseDocument}) => {
   const {t} = useTranslation()
@@ -85,62 +78,11 @@ export const DocumentPerspectiveList = memo(function DocumentPerspectiveList() {
   const setPerspective = useSetPerspective()
   const {params} = usePaneRouter()
   const dateTimeFormat = useDateTimeFormat(DATE_TIME_FORMAT)
-  const {data: releases, loading} = useActiveReleases()
-  const {data: archivedReleases} = useArchivedReleases()
+  const {loading} = useActiveReleases()
   const schema = useSchema()
   const {editState, displayed, documentType, documentId} = useDocumentPane()
-  const {data: documentVersions} = useDocumentVersions({documentId})
   const isCreatingDocument = displayed && !displayed._createdAt
-
-  const filteredReleases: FilterReleases = useMemo(() => {
-    if (!documentVersions) return {notCurrentReleases: [], currentReleases: [], inCreation: null}
-    // Gets the releases ids from the document versions, it means, the releases that the document belongs to
-    const releasesIds = documentVersions.map((id) => getVersionFromId(id))
-    const activeReleases = releases.reduce(
-      (acc: FilterReleases, release) => {
-        const versionDocExists = releasesIds.includes(
-          getReleaseIdFromReleaseDocumentId(release._id),
-        )
-        const releaseId = getReleaseIdFromReleaseDocumentId(release._id)
-        const isCreatingThisVersion =
-          isCreatingDocument &&
-          releaseId === getVersionFromId(displayed._id || '') &&
-          releaseId === selectedReleaseId
-
-        if (isCreatingThisVersion) {
-          acc.inCreation = release
-        } else if (versionDocExists) {
-          acc.currentReleases.push(release)
-        } else {
-          acc.notCurrentReleases.push(release)
-        }
-        return acc
-      },
-      {notCurrentReleases: [], currentReleases: [], inCreation: null},
-    )
-
-    // without historyVersion, version is not in an archived release
-    if (!params?.historyVersion) return activeReleases
-
-    const archivedRelease = archivedReleases.find(
-      (r) => getReleaseIdFromReleaseDocumentId(r._id) === params?.historyVersion,
-    )
-
-    // only for explicitly archived releases; published releases use published perspective
-    if (archivedRelease?.state === 'archived') {
-      activeReleases.currentReleases.push(archivedRelease)
-    }
-
-    return activeReleases
-  }, [
-    archivedReleases,
-    isCreatingDocument,
-    displayed?._id,
-    documentVersions,
-    params?.historyVersion,
-    releases,
-    selectedReleaseId,
-  ])
+  const filteredReleases = useFilteredReleases({displayed, documentId})
 
   const handlePerspectiveChange = useCallback(
     (perspective: Parameters<typeof setPerspective>[0]) => () => {
